gpexpand: skip expanded/broken tables when redistributing
This commit is a cherry-pick of the community PR #15961.
In summary `gpexpand redistribution` earlier used to error out when the table would have been already expanded manually by the user or the table would have been damaged. This commit changes this behaviour so that any expanded/broken tables are skipped without affecting the redistribution for other tables.

Co-authored-by: Dmitry Smal <mialinx@yandex-team.ru>
Co-authored-by: Nihal Jain <jnihal@vmware.com>